requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
Brotli==1.1.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
dbus-python==1.2.16
discord.py==2.1.0
distro==1.4.0
dnspython==2.4.2
frozenlist==1.3.3
idna==2.8
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
multidict==6.0.4
mutagen==1.47.0
mypy-extensions==0.4.3
numpy==1.24.1
openai==0.27.0
opencv-python==4.7.0.68
packaging==23.1
pathspec==0.10.3
platformdirs==2.6.2
pycairo==1.24.0
pycryptodomex==3.18.0
Pygments==2.3.1
PyGObject==3.36.0
pymongo==4.5.0
pyparsing==2.4.6
python-dotenv==0.21.0
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
six==1.14.0
ssh-import-id==5.10
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.25.8
websockets==11.0.3
yarl==1.8.2
yt-dlp==2023.3.3
